Changed state type to float

+++ Source/Collis_Avoid_Sys/Client.h
@@ -30,9 +30,9 @@ public:
 	virtual void Tick(float DeltaTime) override;
 
 	void SendMetadata(int numSensors, bool actionDescrete, int numActions);
-	void SendExperience(TArray<int> currentState, int action, TArray<int> nextState, float reward, bool endGame);
-	void SendExperience(TArray<int> currentState, float throttleAction, float steerAction, TArray<int> nextState, float reward, bool endGame);
-	void Predict(TArray<int> currentState);
+	void SendExperience(TArray<float> currentState, int action, TArray<float> nextState, float reward, bool endGame);
+	void SendExperience(TArray<float> currentState, float throttleAction, float steerAction, TArray<float> nextState, float reward, bool endGame);
+	void Predict(TArray<float> currentState);
 
 	void SetDeepAgent(ADeepAgent* Agent);
 };
